backend/requirements.txt
flask
supabase
python-dotenv
sqlalchemy
psycopg2-binary
PyJWT
pytest
requests
flask-cors
bs4
pydantic
tenacity
bleach
beautifulsoup4
lxml
gunicorn
